// client/Magnus-Laser/src/graphql/types.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};
// ...
export type Bounty = {
  __typename?: 'Bounty';
  ID: Scalars['ID']['output'];
  character: Character;
  crimes: Array<Crime>;
  rep: BountyRep;
  speciality: CrimeType;
  status: BountyStatus;
};

export enum BountyRep {
  BRAGGER = 'BRAGGER',
  LOWPRO = 'LOWPRO'
}

export enum BountyStatus {
  ACTIVE = 'ACTIVE',
  CAPTURED = 'CAPTURED',
  DEAD = 'DEAD'
}

export type BribeCrime = Crime & {
  __typename?: 'BribeCrime';
  crimeType: CrimeType;
  multiplier: Scalars['Int']['output'];
  reward: Scalars['Int']['output'];
  target: BribeTarget;
};

export enum BribeTarget {
  CITIZEN = 'CITIZEN',
  COP = 'COP',
  CORPO = 'CORPO',
  EXEC = 'EXEC',
  GANGER = 'GANGER',
  POLITICIAN = 'POLITICIAN'
}
// ...
export type Character = {
  __typename?: 'Character';
  ID: Scalars['ID']['output'];
  attitude: CharacterAttitude;
  image: Scalars['String']['output'];
  name: Scalars['String']['output'];
  type: CharacterType;
};

export enum CharacterAttitude {
  COLD_AND_PROFESSIONAL = 'COLD_AND_PROFESSIONAL',
  CONSIDERS_A_CREW_MEMBER_A_GOOD_FRIEND = 'CONSIDERS_A_CREW_MEMBER_A_GOOD_FRIEND',
  CONSIDERS_A_CREW_MEMBER_A_PARTNER = 'CONSIDERS_A_CREW_MEMBER_A_PARTNER',
  DOESN_T_TRUST_THE_CREW = 'DOESN_T_TRUST_THE_CREW',
  GENERALLY_DISLIKES_THE_CREW = 'GENERALLY_DISLIKES_THE_CREW',
  NEUTRAL_ATTITUDE_TOWARDS_THE_CREW = 'NEUTRAL_ATTITUDE_TOWARDS_THE_CREW',
  THINKS_THE_CREW_COULD_BE_POTENTIAL_ALLIES = 'THINKS_THE_CREW_COULD_BE_POTENTIAL_ALLIES',
  TRUSTS_IN_THE_CREW_S_DECISIONS = 'TRUSTS_IN_THE_CREW_S_DECISIONS',
  WANTS_A_CREW_MEMBER_DEAD = 'WANTS_A_CREW_MEMBER_DEAD',
  WORSHIPS_A_CREW_MEMBER = 'WORSHIPS_A_CREW_MEMBER'
}

export enum CharacterType {
  AVERAGE_CITIZEN_STREETRAT = 'AVERAGE_CITIZEN_STREETRAT',
  CELEBRITY = 'CELEBRITY',
  CORPORATE_EXEC = 'CORPORATE_EXEC',
  GOVERNMENT_OFFICIAL = 'GOVERNMENT_OFFICIAL',
  MERCENARY_FREELANCER_ASSASSIN = 'MERCENARY_FREELANCER_ASSASSIN',
  NETRUNNER_HACKER = 'NETRUNNER_HACKER',
  NOMAD = 'NOMAD',
  POLICE_LAWMAN = 'POLICE_LAWMAN',
  POLITICIAN = 'POLITICIAN',
  POP_STAR = 'POP_STAR',
  RELIGIOUS_FIGURE = 'RELIGIOUS_FIGURE',
  RIPPERDOC = 'RIPPERDOC',
  TECH = 'TECH',
  TERRORIST = 'TERRORIST'
}

export enum CharacterVerb {
  ESCORT = 'ESCORT',
  INTERROGATE = 'INTERROGATE',
  INTIMIDATE = 'INTIMIDATE',
  INVESTIGATE = 'INVESTIGATE',
  KILL = 'KILL',
  RECRUIT = 'RECRUIT',
  RESCUE = 'RESCUE',
  ROB = 'ROB'
}

export type CharacterVerbWrapper = Verb & {
  __typename?: 'CharacterVerbWrapper';
  value: CharacterVerb;
};

export type ContrabandCrime = Crime & {
  __typename?: 'ContrabandCrime';
  crimeType: CrimeType;
  multiplier: Scalars['Int']['output'];
  reward: Scalars['Int']['output'];
  target: ContrabandTarget;
};

export enum ContrabandTarget {
  COSTLY = 'COSTLY',
  EXPENSIVE = 'EXPENSIVE',
  LUXURY = 'LUXURY',
  PREMIUM = 'PREMIUM',
  SUPER_LUXURY = 'SUPER_LUXURY',
  VERY_EXPENSIVE = 'VERY_EXPENSIVE'
}

export type Crime = {
  crimeType: CrimeType;
  multiplier: Scalars['Int']['output'];
  reward: Scalars['Int']['output'];
};

export enum CrimeType {
  BRIBE = 'BRIBE',
  CONTRABAND = 'CONTRABAND',
  DRUG = 'DRUG',
  MURDER = 'MURDER',
  THEFT = 'THEFT'
}
// ...
export type DrugCrime = Crime & {
  __typename?: 'DrugCrime';
  crimeType: CrimeType;
  multiplier: Scalars['Int']['output'];
  reward: Scalars['Int']['output'];
  target: DrugTarget;
};

export enum DrugTarget {
  BLACK_LACE = 'BLACK_LACE',
  BLUE_GLASS = 'BLUE_GLASS',
  BOOST = 'BOOST',
  SMASH = 'SMASH',
  STIM = 'STIM',
  SYNTHCOKE = 'SYNTHCOKE'
}
// ...
export type MurderCrime = Crime & {
  __typename?: 'MurderCrime';
  crimeType: CrimeType;
  multiplier: Scalars['Int']['output'];
  reward: Scalars['Int']['output'];
  target: MurderTarget;
};

export enum MurderTarget {
  CITIZEN = 'CITIZEN',
  COP = 'COP',
  CORPO = 'CORPO',
  EXEC = 'EXEC',
  GANGER = 'GANGER',
  POLITICIAN = 'POLITICIAN'
}
// ...
export type TheftCrime = Crime & {
  __typename?: 'TheftCrime';
  crimeType: CrimeType;
  multiplier: Scalars['Int']['output'];
  reward: Scalars['Int']['output'];
  target: TheftTarget;
};

export enum TheftTarget {
  MEGA_CORP = 'MEGA_CORP',
  MINI_CORP = 'MINI_CORP',
  OASIS = 'OASIS',
  POOR_SOUL = 'POOR_SOUL',
  PUBLIC_ELEMENT = 'PUBLIC_ELEMENT',
  VENDIT = 'VENDIT'
}